Refactor: Modularize tab components into separate files.
This commit separates the UI and logic for each main tab
(Main Calculation, Room Calculations, Calculation History, Supabase Config)
into its own Python module (main_tab.py, rooms_tab.py, etc.).
Custom Qt widgets have been moved to custom_widgets.py.

HomeUnitCalculator.py has been updated to instantiate these tab
modules and acts as the main application shell.

Known issues:
- A helper function `_clear_layout` is duplicated in main_tab.py and
  rooms_tab.py. I was unable to move it to utils.py.
- I encountered a Qt 'xcb' platform plugin error in the execution environment after Python-level errors were fixed, which blocked full GUI testing.

This refactoring improves code organization and maintainability.
